Added movement or reference frame
Added sliders to move in 3D (not rotate yet) the reference frame. That way, we can calculate the arm lengths for rotations around arbitrary axis.
Also rotated the general geometry so the default reference frame matches the direction of the chair. That is perpendicular to one of the short arms at the top. It used to be perpendicular to the long arm.